TIP-236: Move form event to the namespace form
