+++ src/Pim/Bundle/UserBundle/Form/Type/UserType.php
@@ -6,7 +6,7 @@ use Pim\Bundle\UIBundle\Form\Type\DateType;
 use Pim\Bundle\UserBundle\Doctrine\ORM\Repository\GroupRepository;
 use Pim\Bundle\UserBundle\Doctrine\ORM\Repository\RoleRepository;
 use Pim\Bundle\UserBundle\Entity\UserInterface;
-use Pim\Bundle\UserBundle\Event\UserFormBuilderEvent;
+use Pim\Bundle\UserBundle\Form\Event\UserFormBuilderEvent;
 use Pim\Bundle\UserBundle\Form\Subscriber\UserPreferencesSubscriber;
 use Pim\Bundle\UserBundle\Form\Subscriber\UserSubscriber;
 use Pim\Component\User\Model\Group;
